Create a new file for wall table view cell
